+++ src/emqttc_keepalive.erl
@@ -1,5 +1,5 @@
 %%%-----------------------------------------------------------------------------
-%%% Copyright (c) 2012-2015 eMQTT.IO, All Rights Reserved.
+%%% Copyright (c) 2012-2016 eMQTT.IO, All Rights Reserved.
 %%%
 %%% Permission is hereby granted, free of charge, to any person obtaining a copy
 %%% of this software and associated documentation files (the "Software"), to deal
@@ -19,15 +19,12 @@
 %%% OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 %%% SOFTWARE.
 %%%-----------------------------------------------------------------------------
-%%% @doc
-%%% emqttc socket keepalive.
+%%% @doc emqttc socket keepalive.
 %%%
-%%% @end
+%%% @author Feng Lee <feng@emqtt.io>
 %%%-----------------------------------------------------------------------------
 
 -module(emqttc_keepalive).
-
--author("Feng Lee <feng@emqtt.io>").
 
 -record(keepalive, {socket,
                     stat_name,
@@ -36,23 +33,20 @@
                     timeout_msg,
                     timer_ref}).
 
--opaque keepalive() :: #keepalive{}.
+-opaque keepalive() :: #keepalive{} | undefined.
 
 -export_type([keepalive/0]).
 
 %% API
 -export([new/3, start/1, restart/1, resume/1, cancel/1]).
 
-%%------------------------------------------------------------------------------
 %% @doc Create a KeepAlive.
-%% @end
-%%------------------------------------------------------------------------------
 -spec new({Socket, StatName}, TimeoutSec, TimeoutMsg) -> KeepAlive when
     Socket        :: inet:socket() | ssl:sslsocket(),
     StatName      :: recv_oct | send_oct,
     TimeoutSec    :: non_neg_integer(),
     TimeoutMsg    :: tuple(),
-    KeepAlive     :: keepalive() | undefined.
+    KeepAlive     :: keepalive().
 new({_Socket, _StatName}, 0, _TimeoutMsg) ->
     undefined;
 new({Socket, StatName}, TimeoutSec, TimeoutMsg) when TimeoutSec > 0 ->
@@ -61,35 +55,27 @@ new({Socket, StatName}, TimeoutSec, TimeoutMsg) when TimeoutSec > 0 ->
                timeout_sec = TimeoutSec,
                timeout_msg = TimeoutMsg}.
 
-%%------------------------------------------------------------------------------
 %% @doc Start KeepAlive
-%% @end
-%%------------------------------------------------------------------------------
--spec start(KeepAlive) -> KeepAlive when
-    KeepAlive   :: keepalive() | undefined.
+-spec start(keepalive()) -> {ok, keepalive()} | {error, any()}.
 start(undefined) ->
-    undefined;
+    {ok, undefined};
 start(KeepAlive = #keepalive{socket = Socket, stat_name = StatName, 
                              timeout_sec = TimeoutSec, 
                              timeout_msg = TimeoutMsg}) ->
-    {ok, [{StatName, StatVal}]} = emqttc_socket:getstat(Socket, [StatName]),
-    Ref = erlang:send_after(TimeoutSec*1000, self(), TimeoutMsg),
-    KeepAlive#keepalive{stat_val = StatVal, timer_ref = Ref}.
+    case emqttc_socket:getstat(Socket, [StatName]) of
+        {ok, [{StatName, StatVal}]} ->
+            Ref = erlang:send_after(TimeoutSec*1000, self(), TimeoutMsg),
+            {ok, KeepAlive#keepalive{stat_val = StatVal, timer_ref = Ref}};
+        {error, Error} ->
+            {error, Error}
+    end.
 
-%%------------------------------------------------------------------------------
 %% @doc Restart KeepAlive
-%% @end
-%%------------------------------------------------------------------------------
--spec restart(KeepAlive) -> KeepAlive when
-    KeepAlive   :: keepalive() | undefined.
+-spec restart(keepalive()) -> {ok, keepalive()} | {error, any()}.
 restart(KeepAlive) -> start(KeepAlive).
 
-%%------------------------------------------------------------------------------
 %% @doc Resume KeepAlive, called when timeout.
-%% @end
-%%------------------------------------------------------------------------------
--spec resume(KeepAlive) -> timeout | {resumed, KeepAlive} when
-    KeepAlive  :: keepalive() | undefined.
+-spec resume(keepalive()) -> timeout | {resumed, keepalive()} | {error, any()}.
 resume(undefined) -> undefined;
 resume(KeepAlive = #keepalive{socket      = Socket,
                               stat_name   = StatName,
@@ -97,21 +83,22 @@ resume(KeepAlive = #keepalive{socket      = Socket,
                               timeout_sec = TimeoutSec,
                               timeout_msg = TimeoutMsg,
                               timer_ref   = Ref}) ->
-    {ok, [{StatName, NewStatVal}]} = emqttc_socket:getstat(Socket, [StatName]),
-    if
-        NewStatVal =:= StatVal ->
-            {error, timeout};
-        true ->
-            cancel(Ref), %need?
-            NewRef = erlang:send_after(TimeoutSec*1000, self(), TimeoutMsg),
-            {resumed, KeepAlive#keepalive{stat_val = NewStatVal, timer_ref = NewRef}}
+    case emqttc_socket:getstat(Socket, [StatName]) of
+        {ok, [{StatName, NewStatVal}]} ->
+            if
+                NewStatVal =:= StatVal ->
+                    timeout;
+                true ->
+                    cancel(Ref), %need?
+                    NewRef = erlang:send_after(TimeoutSec*1000, self(), TimeoutMsg),
+                    {resumed, KeepAlive#keepalive{stat_val = NewStatVal, timer_ref = NewRef}}
+            end;
+        {error, Error} ->
+            {error, Error}
     end.
 
-%%------------------------------------------------------------------------------
 %% @doc Cancel KeepAlive.
-%% @end
-%%------------------------------------------------------------------------------
--spec cancel(keepalive() | undefined | reference()) -> any().
+-spec cancel(keepalive() | reference()) -> any().
 cancel(undefined) ->
     ok;
 cancel(#keepalive{timer_ref = Ref}) ->